* avr.c (avr_output_function_prologue, 	avr_output_function_epilogue, print_operand): Fix format specifier 	warnings. 	(init_cumulative_args): Mark parameter with ATTRIBUTE_UNUSED. 	* avr.h (FUNCTION_VALUE_REGNO_P): Fix signed/unsigned warnings.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@66895 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/config/avr/avr.h
@@ -1224,7 +1224,7 @@ extern int avr_reg_order[];
    data types, because none of the library functions returns such
    types.  */
 
-#define FUNCTION_VALUE_REGNO_P(N) ((N) == RET_REGISTER)
+#define FUNCTION_VALUE_REGNO_P(N) ((int) (N) == RET_REGISTER)
 /* A C expression that is nonzero if REGNO is the number of a hard
    register in which the values of called function may come back.
